refactor: 서버컴포넌트로 변경 (#254) - 타입, 상수 분리 - useGachaGame 훅으로 분리 - GachaGame 컴포넌트 클라이언트 컴포넌트로 변경 - 앱 라우팅하는 페이지를 서버 컴포넌트로 변경 - views/GachaGame.tsx 파일 삭제
