+++ apps/web/app/gacha/page.tsx
@@ -1,0 +1,7 @@
+import GachaGameClient from '@/components/game/GachaGameClient';
+
+const GotchaPage = () => {
+  return <GachaGameClient />;
+};
+
+export default GotchaPage;
