--- apps/web/app/gotcha/page.tsx
@@ -1,10 +1,0 @@
-import GachaGame from '@/views/GachaGame';
-const page = () => {
-  return (
-    <div>
-      <GachaGame />
-    </div>
-  );
-};
-
-export default page;
