+++ apps/web/src/components/game/GachaGameClient.tsx
@@ -1,169 +1,25 @@
 'use client';
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 
 import Image from 'next/image';
 
 import { AnimatePresence, motion } from 'motion/react';
 
-// 타입 정의
-interface Ball {
-  id: number;
-  x: number;
-  y: number;
-  rotate: number;
-  hue: number;
-  size: number;
-}
+import { useGachaGame } from '@/hooks/common/useGachaGame';
 
-interface Prize {
-  image: string;
-  title: string;
-}
-
-interface ConfettiParticle {
-  id: number;
-  x: number;
-  y: number;
-  speedX: number;
-  speedY: number;
-  size: number;
-  color: string;
-  rx: number;
-  ry: number;
-  rz: number;
-  rs: number;
-}
-
-const prizes: Prize[] = [
-  {
-    image: 'https://assets.codepen.io/2509128/prize1.png',
-    title: '토끼 인형',
-  },
-  {
-    image: 'https://assets.codepen.io/2509128/prize2.png',
-    title: '테디베어',
-  },
-  {
-    image: 'https://assets.codepen.io/2509128/prize3.png',
-    title: '북극곰',
-  },
-  {
-    image: 'https://assets.codepen.io/2509128/prize4.png',
-    title: '북극곰 신부',
-  },
-];
-
-// 갓챠 게임 컴포넌트
-const GachaGame: React.FC = () => {
-  const [gameState, setGameState] = useState<'init' | 'ready' | 'playing' | 'complete'>('init');
-  const [balls, setBalls] = useState<Ball[]>([]);
-  const [selectedPrize, setSelectedPrize] = useState<Prize | null>(null);
-  const [showConfetti, setShowConfetti] = useState(false);
-  const [confettiParticles, setConfettiParticles] = useState<ConfettiParticle[]>([]);
-  const [isJittering, setIsJittering] = useState(false);
-  const [showHint, setShowHint] = useState(false);
-  const [showHint2, setShowHint2] = useState(false);
-
-  // 볼 생성
-  const createBalls = (): Ball[] => {
-    const ballPositions = [
-      { x: 0.5, y: 0.6 },
-      { x: 0, y: 0.68 },
-      { x: 0.22, y: 0.65 },
-      { x: 0.7, y: 0.63 },
-      { x: 0.96, y: 0.66 },
-      { x: 0.75, y: 0.79 },
-      { x: 0.5, y: 0.8 },
-      { x: 0.9, y: 0.81 }, // 상품 볼
-      { x: 0, y: 0.82 },
-      { x: 1, y: 0.9 },
-      { x: 0.25, y: 0.85 },
-      { x: 0.9, y: 1 },
-      { x: 0.4, y: 1 },
-      { x: 0.65, y: 1 },
-      { x: 0.09, y: 1 },
-    ];
-
-    return ballPositions.map((pos, index) => ({
-      id: index,
-      x: pos.x,
-      y: pos.y,
-      rotate: Math.floor(Math.random() * 360),
-      hue: Math.floor(Math.random() * 360),
-      size: 8,
-    }));
-  };
-
-  // 컨페티 생성
-  const createConfetti = () => {
-    const particles: ConfettiParticle[] = [];
-    for (let i = 0; i < 128; i++) {
-      particles.push({
-        id: i,
-        x: 50 + (Math.random() * 12 - 6),
-        y: 50 + (Math.random() * 12 - 6),
-        speedX: Math.random() * 1.5 - 0.75,
-        speedY: -0.5 + Math.random() * 2 - 1,
-        size: 8 + Math.random() * 8 - 4,
-        color: `hsl(${Math.random() * 360}deg, 80%, 60%)`,
-        rx: Math.random() * 2 - 1,
-        ry: Math.random() * 2 - 1,
-        rz: Math.random() * 2 - 1,
-        rs: Math.random() * 2 + 0.5,
-      });
-    }
-    setConfettiParticles(particles);
-    setShowConfetti(true);
-
-    // 3초 후 컨페티 제거
-    setTimeout(() => {
-      setShowConfetti(false);
-      setConfettiParticles([]);
-    }, 3000);
-  };
-
-  // 게임 초기화
-  useEffect(() => {
-    // 랜덤 상품 선택 함수를 useEffect 내부로 이동
-    const getRandomPrize = (): Prize => {
-      return prizes[Math.floor(Math.random() * prizes.length)] as Prize;
-    };
-
-    const newBalls = createBalls();
-    setBalls(newBalls);
-    setSelectedPrize(getRandomPrize());
-
-    setTimeout(() => {
-      setGameState('ready');
-      setTimeout(() => setShowHint(true), 2000);
-    }, 1000);
-  }, []);
-
-  // 핸들 클릭
-  const handleClick = async () => {
-    if (gameState !== 'ready') return;
-
-    setGameState('playing');
-    setShowHint(false);
-    setIsJittering(true);
-
-    // 2초 후 지터링 중지하고 볼 떨어뜨리기
-    setTimeout(() => {
-      setIsJittering(false);
-      setTimeout(() => {
-        setShowHint2(true);
-      }, 2000);
-    }, 2000);
-  };
-
-  // 상품 볼 클릭
-  const handlePrizeBallClick = () => {
-    if (gameState !== 'playing') return;
-
-    setShowHint2(false);
-    setGameState('complete');
-    createConfetti();
-  };
+const GachaGameClient: React.FC = () => {
+  const {
+    gameState,
+    balls,
+    selectedPrize,
+    showConfetti,
+    confettiParticles,
+    isJittering,
+    showHint,
+    showHint2,
+    handleClick,
+    handlePrizeBallClick,
+  } = useGachaGame();
 
   return (
     <div className="relative h-screen w-full overflow-hidden bg-gray-600">
@@ -501,7 +357,6 @@ const GachaGame: React.FC = () => {
               <motion.img
                 src="https://assets.codepen.io/2509128/shine.png"
                 className="absolute h-screen"
-                // animate={{ rotate: 360 }}
                 transition={{
                   duration: 5,
                   repeat: Infinity,
@@ -554,4 +409,4 @@ const GachaGame: React.FC = () => {
   );
 };
 
-export default GachaGame;
+export default GachaGameClient;
