+++ apps/web/src/lib/constants/game.ts
@@ -1,0 +1,20 @@
+import { Prize } from '@/types/game';
+
+export const prizes: Prize[] = [
+  {
+    image: 'https://assets.codepen.io/2509128/prize1.png',
+    title: '토끼 인형',
+  },
+  {
+    image: 'https://assets.codepen.io/2509128/prize2.png',
+    title: '테디베어',
+  },
+  {
+    image: 'https://assets.codepen.io/2509128/prize3.png',
+    title: '북극곰',
+  },
+  {
+    image: 'https://assets.codepen.io/2509128/prize4.png',
+    title: '북극곰 신부',
+  },
+];
