+++ apps/web/src/lib/types/index.ts
@@ -1,2 +1,3 @@
 // TypeScript 타입 정의들
+export * from './game';
 export * from './users';
